refactor: remove task list infrastructure in favor of tags (#89)
Task lists were the original organizational mechanism but tags have
fully replaced them. This removes all task list code across the stack:

- Delete TaskList model, controller, policy, form request, resource
- Remove ManageTaskLists MCP tool and list_id from ManageTasks
- Remove /task-lists API routes and task_list_id from task CRUD
- Delete TaskListsView and TaskListDetailView frontend components
- Rename TaskListStep.vue to TagsStep.vue in onboarding wizard
- Remove taskLists relationship from Family model
- Add migration to drop task_list_id column and task_lists table

8 files deleted, 12 files modified. Frontend build passes clean.

Closes #89
