--- app/Http/Requests/Task/StoreTaskListRequest.php
@@ -1,33 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Task;
-
-use Illuminate\Foundation\Http\FormRequest;
-
-class StoreTaskListRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array|string>
-     */
-    public function rules(): array
-    {
-        return [
-            'name' => 'required|string|max:255',
-            'description' => 'nullable|string',
-            'icon' => 'nullable|string',
-            'color' => 'nullable|string|max:20',
-        ];
-    }
-}
