+++ app/Http/Requests/Task/UpdateTaskRequest.php
@@ -24,7 +24,6 @@ class UpdateTaskRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'task_list_id' => 'sometimes|nullable|exists:task_lists,id',
             'title' => 'sometimes|string|max:255',
             'tag_ids' => 'nullable|array',
             'tag_ids.*' => 'exists:tags,id',
