--- app/Models/TaskList.php
@@ -1,57 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Concerns\HasUuids;
-use Illuminate\Database\Eloquent\Factories\HasFactory;
-use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
-use Illuminate\Database\Eloquent\Relations\HasMany;
-
-class TaskList extends Model
-{
-    use HasFactory, HasUuids;
-
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array<int, string>
-     */
-    protected $fillable = [
-        'family_id',
-        'name',
-        'description',
-        'icon',
-        'color',
-        'sort_order',
-        'is_default',
-    ];
-
-    /**
-     * Get the attributes that should be cast.
-     *
-     * @return array<string, string>
-     */
-    protected function casts(): array
-    {
-        return [
-            'is_default' => 'boolean',
-        ];
-    }
-
-    /**
-     * TaskList belongs to a Family.
-     */
-    public function family(): BelongsTo
-    {
-        return $this->belongsTo(Family::class);
-    }
-
-    /**
-     * TaskList has many Tasks.
-     */
-    public function tasks(): HasMany
-    {
-        return $this->hasMany(Task::class);
-    }
-}
